+++ src/ScrapBook/Collecter.hs
@@ -8,6 +8,7 @@ import           Data.Extensible
 import           Data.Extensible.Effect.Default (EitherDef, runEitherDef)
 import           Data.Extensible.Effect.Logger
 import           Data.Text                      (Text)
+import           Data.Yaml                      (ParseException)
 import           Network.HTTP.Req               (HttpException)
 
 type Collecter = Eff
@@ -20,9 +21,13 @@ data CollectError
   = FetchException (Either HttpException Text)
   | WriteException Text
   | CollectException Text
+  | YamlParseException ParseException
   deriving (Show, Eq)
 
 instance Eq HttpException where
+  a == b = show a == show b
+
+instance Eq ParseException where
   a == b = show a == show b
 
 collect :: Collecter a -> IO (Either CollectError a)