+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Feat: version option
 - Feat: RSS 2.0
+- Fix: remove namespace in xml tag
 
 ## Alpha