+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Feat: version option
 - Feat: RSS 2.0
 - Fix: remove namespace in xml tag
+- Feat: summary
 
 ## Alpha