+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Feat: RSS 2.0
 - Fix: remove namespace in xml tag
 - Feat: summary
+- Feat: mltiple input files
 
 ## Alpha